requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-filter==2.0.0
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
Markdown==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
requests==2.21.0
requirements==0.1
schedule==0.6.0
urllib3==1.24.1
